rename jsx to jsx-a11y
