+++ src/index.ts
@@ -2,7 +2,7 @@ import base from "./base.js";
 import stylistic from "./stylistic.js";
 import typescript from "./typescript.js";
 import react from "./react.js";
-import jsx from "./jsx.js";
+import jsxAccessibility from "./jsx-a11y.js";
 import importPlugin from "./import.js";
 import unusedImports from "./unused-imports.js";
 
@@ -11,7 +11,7 @@ export default [
   ...stylistic,
   ...typescript,
   ...react,
-  ...jsx,
+  ...jsxAccessibility,
   ...importPlugin,
   ...unusedImports,
 ];
